requirements.txt
hockey_scraper
pandas
pytest
pytest-cov
requests
beautifulsoup4
luigi
scikit-learn
pyyaml